+++ doc/changelog.rst
@@ -34,6 +34,7 @@ Localization
 Admin
 ~~~~~
 
+- Show list of shipments in order view
 - Fix customer, creator, and ordered by links on order detail page
 - Prevent picotable from reloading after every change
 - Add ability to copy category visibility settings to products